requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
PyJWT==2.9.0
PyMuPDF==1.26.0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
setuptools==78.1.0
urllib3==2.2.2
Werkzeug==3.0.4
yt-dlp==2025.4.30
zope.event==5.0
zope.interface==7.2
